% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis_expansion_function.R
\name{get_basis_expansion}
\alias{get_basis_expansion}
\title{Construct basis expansion matrices.}
\usage{
get_basis_expansion(df, num.knots, bases, interactions, knots.list = NULL)
}
\arguments{
\item{df}{data frame for which to expand. Must already contain an intercept in column 1, and all covariates should be scaled between 0 and 1.}

\item{num.knots}{a vector containing the number of knots to use for each predictor.}

\item{bases}{a vector containing the bases to use for each expansion. Cubic (\code{cr}) and cyclic (\code{cc}) radial splines are supported.}

\item{interactions}{logical: should two-way interactions (tensor products) be computed?}

\item{knots.list}{optional list containing knots to use. Useful for predictions. If present, num.knots is ignored.}
}
\value{
A list with components:
\itemize{
 \item{\code{C.full}}{ a matrix containing all basis expansions, with the intercept.}
 \item{\code{partition}}{ a numeric vector corresponding to group identification. \code{0} corresponds to the intercept.}
 \item{\code{nvar, pc, pz, px}}{ scalars giving the number of total predictors (including 2-way interactions as individual predictors, if \code{interactions == TRUE}),
 total dimension of \code{C.full}, and the dimensions of the linear and basis expansion terms, respectively.}
 \item{\code{knots.list}}{ a list of knots used for expansion. Useful for creating the same basis expansion on an independent data set (for, say, validation).}
}
}
\description{
This function was created to avoid repeating it over and over in my research. The penalty for each marginal term is \eqn{\int \|f''(x)\|^2 dx}.
}

