% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlot.R
\name{multiplot}
\alias{multiplot}
\title{Plot multiple ggplot2s in one page.}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{plot objects for plotting.}

\item{plotlist}{alternative input of plots, as list.}

\item{cols}{number of columns in viewing window.}

\item{layout}{layout matrix for plots.}
}
\description{
Stolen shamelessly from http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{

rm(list=ls())

library(ggplot2)
library(hunterr)

# This example uses the ChickWeight dataset, which comes with ggplot2
# First plot
p1 <- ggplot(ChickWeight, aes(x=Time, y=weight, colour=Diet, group=Chick)) +
 geom_line() +
 ggtitle("Growth curve for individual chicks")

# Second plot
p2 <- ggplot(ChickWeight, aes(x=Time, y=weight, colour=Diet)) +
 geom_point(alpha=.3) +
 geom_smooth(alpha=.2, size=1) +
 ggtitle("Fitted growth curve per diet")

# Third plot
p3 <- ggplot(subset(ChickWeight, Time==21), aes(x=weight, colour=Diet)) +
 geom_density() +
 ggtitle("Final weight, by diet")

# Fourth plot
p4 <- ggplot(subset(ChickWeight, Time==21), aes(x=weight, fill=Diet)) +
 geom_histogram(colour="black", binwidth=50) +
 facet_grid(Diet ~ .) +
 ggtitle("Final weight, by diet") +
 theme(legend.position="none")        # No legend (redundant in this graph)

multiplot(p1, p2, p3, p4, cols=2)
multiplot(p1, p2, p3, p4, layout = matrix(c(1,2,3,4,4,4), nrow=2, byrow=TRUE))

}

